modules/molmo/requirements.txt
transformers
Pillow
fastapi
uvicorn
torch
python-multipart
requests
SpeechRecognition
pyttsx3
gunicorn
torchvision
einops
accelerate
tensorflow